Allow tree generation to cross chunks (decoration vs population)
Previously chunks were 'populated' with trees, editing the chunk voxel
data directly, before it was sent to the engine for rendering, but this
limited feature generation. Add a new 'decorate' phase and message, where
voxel-land generates a list of world global coordinates and values, then
sends them to voxel-engine using setBlock() (may need to change later if
is too slow) -- this allows the generated features to span multiple chunks,
as long as they are loaded. Use new treetype:2 space-filling tree as a demo.